+++ resolwe/observers/management/__init__.py
@@ -1,0 +1,10 @@
+""".. Ignore pydocstyle D400.
+
+====================
+Observers Management
+====================
+
+.. automodule:: resolwe.observers.management.commands.register
+    :members:
+
+"""